Refactor scheduler.py and AllocatorID
