+++ gllm/worker.py
@@ -10,7 +10,7 @@ from gllm.input_data import InputData
 from gllm.model_runner import ModelRunner
 from gllm.zmq_comm import zmqComm
 from gllm.pp_scheduler import PPScheduler
-from gllm.scheduler import IPCPackage
+from gllm.frontend_scheduler import IPCPackage
 from gllm.dist_utils import (init_dist, send_pp_data, recv_pp_data, 
                              get_rank, get_world_size, is_last_pp_rank,
                              get_pp_size, get_next_pp_rank, get_last_pp_rank)
